Rename logging helper and revert usage
