+++ backend/api/assistant/routes.py
@@ -9,7 +9,7 @@ from backend.api.assistant.models import (
     AssistantDeletion
 )
 from backend.util.auth_utils import validate_user, TokenData
-from backend.util.logging import SetupLogging
+from backend.util.log_setup import SetupLogging
 from backend.util.domain_utils import extract_subdomain
 from backend.util.database import get_async_session 
 from sqlalchemy.ext.asyncio import AsyncSession
