+++ backend/api/auth/repository.py
@@ -7,7 +7,7 @@ from sqlalchemy.sql import func
 
 from backend.api.auth.user_schema import UserORM
 from backend.api.auth.models import User
-from backend.util.logging import SetupLogging
+from backend.util.log_setup import SetupLogging
 
 logger = SetupLogging()
