+++ backend/api/auth/routes.py
@@ -5,7 +5,7 @@ from typing import Optional
 from sqlalchemy.ext.asyncio import AsyncSession
 from backend.util.database import get_async_session
 
-from backend.util.logging import SetupLogging
+from backend.util.log_setup import SetupLogging
 from backend.util.auth_utils import validate_user, TokenData, create_access_token
 from backend.util.domain_utils import extract_subdomain
 from backend.util.token_blacklist import add_token_to_blacklist
