+++ backend/api/chat/routes.py
@@ -4,7 +4,7 @@ from fastapi.responses import StreamingResponse
 from backend.api.chat.models import ChatRequest, FeedbackRequest, MessageList
 from backend.api.chat.services import ChatService
 from backend.util.auth_utils import validate_user, TokenData
-from backend.util.logging import SetupLogging
+from backend.util.log_setup import SetupLogging
 from backend.util.database import get_async_session
 from sqlalchemy.ext.asyncio import AsyncSession
