+++ backend/api/chat/services.py
@@ -20,7 +20,7 @@ from backend.api.session.models import UserSession
 from backend.api.session.repository import SessionRepository
 from backend.api.assistant.repository import AssistantRepository
 from backend.api.kbase.repository import KbaseRepository
-from backend.util.logging import SetupLogging
+from backend.util.log_setup import SetupLogging
 from backend.util.auth_utils import TokenData
 
 logger = SetupLogging()
