+++ backend/api/session/routes.py
@@ -6,7 +6,7 @@ from backend.api.session.models import (
 )
 from backend.api.session.services import SessionService
 from backend.util.auth_utils import validate_user, TokenData
-from backend.util.logging import SetupLogging
+from backend.util.log_setup import SetupLogging
 from backend.util.database import get_async_session
 from sqlalchemy.ext.asyncio import AsyncSession
