+++ backend/app.py
@@ -28,7 +28,7 @@ from backend.lib.exceptions import (
     DatabaseError
 )
 from backend.util.config import get_config
-from backend.util.logging import SetupLogging
+from backend.util.log_setup import SetupLogging
 from fastapi.security import OAuth2PasswordBearer
 from passlib.context import CryptContext
 import os
